pyproject.toml
[project]
name = "agentic-data-contracts"
version = "0.10.0"
description = "YAML-first, domain-driven data governance for AI agents"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Qing", email = "qingye779@gmail.com" }]
license = { text = "MIT" }
keywords = [
    "ai-agents",
    "data-contracts",
    "data-governance",
    "domain-driven",
    "sql-validation",
    "llm",
    "claude",
    "semantic-layer",
    "analytics",
    "dbt",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
]
dependencies = [
    "sqlglot>=23.0",
    "pydantic>=2.0",
    "pyyaml>=6.0.3",
    "thefuzz>=0.22.1",
]

[project.urls]
Homepage = "https://github.com/flyersworder/agentic-data-contracts"
Repository = "https://github.com/flyersworder/agentic-data-contracts"
Issues = "https://github.com/flyersworder/agentic-data-contracts/issues"
Documentation = "https://github.com/flyersworder/agentic-data-contracts/blob/main/docs/architecture.md"

[project.optional-dependencies]
agent-sdk = ["claude-agent-sdk>=0.1.58"]
agent-contracts = ["ai-agent-contracts>=0.3.0"]
bigquery = ["google-cloud-bigquery"]
snowflake = ["snowflake-connector-python"]
postgres = ["psycopg2-binary"]
duckdb = ["duckdb"]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "duckdb",
]
all = [
    "agentic-data-contracts[agent-sdk,agent-contracts,bigquery,snowflake,postgres,duckdb]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_data_contracts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.rules]
unresolved-import = "ignore"
